added courses to schedules

+++ static/schedule.js
@@ -1,7 +1,5 @@
 document.getElementById('search-courses').addEventListener('click', courseSearch);
 document.getElementById('schedule-create').addEventListener('click', makeSchedule);
-document.getElementById('schedules').addEventListener('onchange', displaySchedule);
-document.getElementById('schedule-display').addEventListener('click', displaySchedule);
 document.getElementById('schedule-delete').addEventListener('click', deleteSchedule);
 document.getElementById('schedule-delete-all').addEventListener('click', deleteAllSchedules);
 document.getElementById('add-course').addEventListener('click', addCourse);
@@ -9,6 +7,10 @@ document.getElementById('add-course').addEventListener('click', addCourse);
 
 getCourses();
 getSchedule();
+//displaySchedule();
+
+var newCourses = [];
+const scheduleNames = [];
 
 function getCourses(){
     clearArea("courses");
@@ -16,7 +18,6 @@ function getCourses(){
     fetch("/api/courses")
     .then(res => res.json()
     .then(data => {
-        console.log(data);
         const l = document.getElementById('courses');
         data.forEach(e => {
             const table = document.createElement('table');
@@ -33,9 +34,12 @@ function getCourses(){
             row.appendChild(cell2)
             table.appendChild(row)
             l.appendChild(table);
-
-            //const br = document.createElement('br');
-            //l.appendChild(br);
+            if(e.course_info[0].ssr_component=="LAB"){
+                table.id = "lab";
+            }
+            if(e.course_info[0].ssr_component=="TUT"){
+                table.id = "tut";
+            }
         })
     })
     )
@@ -47,6 +51,7 @@ function getSchedule(){
 
     fetch('/api/schedule').then(res => res.json())
     .then(data => {
+        scheduleNames.push(data);
         const l = document.getElementById('schedules');
         var empty = true;
 
@@ -71,16 +76,29 @@ function getSchedule(){
 
 function makeSchedule(){
     const l = document.getElementById('schedule-name').value;
-    fetch('/api/schedule/'+ l, {
+    if(!(scheduleNames[0].includes(document.getElementById('schedule-name').value)) && document.getElementById('schedule-name').value!=="" ){
+        scheduleNames[0].push(document.getElementById('schedule-name').value);
+        fetch('/api/schedule/'+ l, {
+            method: 'PUT',
+        })
+        getSchedule();
+        displaySchedule()
+    }
+    else{
+        alert("schedule already exists or nothing was entered");
+    }
+}
+
+function updateSchedule(name){
+    const l = document.getElementById('schedules').value;
+    fetch('/api/schedule/'+ l+'/'+newCourses, {
         method: 'PUT',
     })
-    .then(res => console.log(res))
-    getSchedule();
+    displaySchedule()
 }
 
 function displaySchedule(){
     clearArea('schedule-area');
-    
     const l = document.getElementById('schedules').value;
     fetch('/api/schedule/'+ l, {
         method: 'GET',
@@ -88,22 +106,37 @@ function displaySchedule(){
     .then(res =>  res.json())
     .then(data => {
         const l = document.getElementById('schedule-area');
-        const option = document.createElement('p');
-        option.appendChild(document.createTextNode(data));
-        l.appendChild(option);
-        //newCourses.push(data);
+        const ol1 = document.createElement('ol');
+        newCourses = [];
+        if(data!=""){
+            newCourses.push(data);
+        }
+        //console.log(newCourses[0][0].split(","));
+        for(i=0;i<newCourses[0][0].split(",").length;i+=2){
+            const option = document.createElement('p');
+            option.appendChild(document.createTextNode(newCourses[0][0].split(",")[0]+", "+newCourses[0][0].split(",")[1]));
+            ol1.appendChild(option)
+        }
+
+        l.appendChild(ol1);
     });
 }
 
 
 function deleteSchedule(){
-    const l = document.getElementById('schedules');
-    console.log(l.value)
-    fetch('/api/schedule/'+ l.value, {
-        method: 'DELETE',
-    })
-    .then(res => console.log(res))
+    const l = document.getElementById('schedules').value;
+    if(document.getElementById('schedules').value!="No Schedules"){
+        fetch('/api/schedule/'+ l, {
+            method: 'DELETE',
+        })
+        .then(res => console.log(res))
+    }
+    else{
+        console.log("no schedules to delete")
+    }
+
     getSchedule();
+    displaySchedule()
 }
 
 function deleteAllSchedules(){
@@ -112,33 +145,35 @@ function deleteAllSchedules(){
     fetch('/api/schedule/' , {
         method: 'DELETE',
     })
-    .then(res => console.log(res))
     getSchedule();
+    displaySchedule()
 }
 
 function addCourse(){
-    /*
-    fetch("/api/courses")
-    .then(res => res.json()
-    .then(data => {
-        const l = document.getElementById("schedule-area");
-        if(document.getElementById('courseSubject').value!=="" && document.getElementById('courseCatalog_nbr').value!==""){
-            data.forEach(e => {
-                if(document.getElementById('courseSubject').value==e.subject && document.getElementById('courseCatalog_nbr').value==e.catalog_nbr){
-                    newCourses.push(e.subject,e.catalog_nbr);
-                    updateSchedule(newCourses);
-
-                }
-            })
-        }
-        else{
-            const error = document.createElement('li');
-            error.appendChild(document.createTextNode("No courses found"))
-            l.appendChild(error);
-        }
-    })
-    )*/
-    console.log("pls")
+    if(document.getElementById('schedules').value!="No Schedules"){
+        fetch("/api/courses")
+        .then(res => res.json()
+        .then(data => {
+            const l = document.getElementById("schedule-area");
+            if(document.getElementById('courseSubject').value!=="" && document.getElementById('courseCatalog_nbr').value!==""){
+                data.forEach(e => {
+                    if(document.getElementById('courseSubject').value==e.subject && document.getElementById('courseCatalog_nbr').value==e.catalog_nbr){
+                        newCourses.push(e.subject,e.catalog_nbr)
+                        updateSchedule()
+                    }
+                })
+            }
+            else{
+                const error = document.createElement('li');
+                error.appendChild(document.createTextNode("No courses found"))
+                l.appendChild(error);
+            }
+        })
+        )
+    }
+    else{
+        console.log("no schedule")
+    }
 }
 
 
@@ -149,7 +184,6 @@ function clearArea(area){
 }
 
 function courseSearch(){
-    getSchedule()
     var found = false;
 
     clearArea("courses");
@@ -157,15 +191,27 @@ function courseSearch(){
     fetch("/api/courses")
     .then(res => res.json()
     .then(data => {
-        console.log(data);
         const l = document.getElementById('courses');
+        if(document.getElementById('subject').value=="" && document.getElementById('catalog_nbr').value=="" && document.getElementById('ssr_component').value==""){
+            getCourses();
+        }
         if(document.getElementById('subject').value!=="" && document.getElementById('catalog_nbr').value=="" && document.getElementById('ssr_component').value==""){
             data.forEach(e => {
                 if(document.getElementById('subject').value==e.subject){
-                    const course = document.createElement('li');
-                    course.appendChild(document.createTextNode(`${e.catalog_nbr}`))
-                    l.appendChild(course);
+                    const table = document.createElement('table');
+                    const row = document.createElement('tr');
+                    const cell = document.createElement('td');
+                    cell.appendChild(document.createTextNode(`${e.catalog_nbr}`));
+                    row.appendChild(cell);
+                    table.appendChild(row);
+                    l.appendChild(table);
                     found = true;
+                    if(e.course_info[0].ssr_component=="LAB"){
+                        table.id = "lab";
+                    }
+                    if(e.course_info[0].ssr_component=="TUT"){
+                        table.id = "tut";
+                    }
                 }
             })
         }
@@ -189,8 +235,9 @@ function courseSearch(){
                 }
             })
         }
-        if(found == false){
+        if(found == false && !(document.getElementById('subject').value=="" && document.getElementById('catalog_nbr').value=="" && document.getElementById('ssr_component').value=="")){
             const error = document.createElement('li');
+            alert("no courses found");
             error.appendChild(document.createTextNode("No courses found"))
             l.appendChild(error);
         }
